+++ src/rfc3987_syntax2/__init__.py
@@ -2,7 +2,11 @@
 # Copyright (c) 2026 Jan Kowalleck - modifications and maintenance
 # SPDX-License-Identifier: MIT
 
-from typing import Any
+"""
+Re-export of :mod:`.syntax_helpers`
+"""
+
+from typing import Any, TYPE_CHECKING
 
 from . import syntax_helpers as _syntax_helpers
 
@@ -10,6 +14,9 @@ __all__ = _syntax_helpers.__all__
 
 _EXPORTS_SH = frozenset(_syntax_helpers.__all__)
 
+if TYPE_CHECKING:
+    from .syntax_helpers import *
+# cannot simply re-export - need to respect lazy-loading
 
 def __getattr__(name: str) -> Any:
     if name in _EXPORTS_SH: